+++ crates/crabe_decision/src/strategy/formations.rs
@@ -1,3 +1,7 @@
 //KickOffAlly set ups the kick of when in favor of the kick off
-mod kick_off;
-pub use self::kick_off::KickOffAlly;
+mod kick_off_ally;
+pub use self::kick_off_ally::KickOffAlly;
+
+//KickOffEnemy set ups the kick of when not in favor of the kick off
+mod kick_off_enemy;
+pub use self::kick_off_enemy::KickOffEnemy;
