+++ crates/crabe_decision/src/strategy/formations/kick_off_enemy.rs
@@ -6,25 +6,25 @@ use crabe_framework::data::world::World;
 use nalgebra::Point2;
 use std::f64::consts::PI;
 
-/// The KickOffAlly struct represents a strategy that commands the team to set in the KickOffAlly formation
-/// It is used when the team is in favor of the kick-off
+/// The KickOffEnemy struct represents a strategy that commands the team to set in the KickOffEnemy formation
+/// It is used when the team is not in favor of the kick-off
 #[derive(Default)]
-pub struct KickOffAlly {
+pub struct KickOffEnemy {
     /// The id of the robot to move.
     id: u8,
 }
 
-impl KickOffAlly {
-    /// Creates a new KickOffAlly instance with the desired robot id.
+impl KickOffEnemy {
+    /// Creates a new KickOffEnemy instance with the desired robot id.
     pub fn new(id: u8) -> Self {
         Self { id }
     }
 }
 
-impl Strategy for KickOffAlly {
-    /// Executes the KickOffAlly strategy.
+impl Strategy for KickOffEnemy {
+    /// Executes the KickOffEnemy strategy.
     ///
-    /// This strategy commands all the robots to move in position for
+    /// This strategy commands all the robots to move in position for kick-off when not in favor
     ///
     /// # Arguments
     ///
@@ -42,12 +42,13 @@ impl Strategy for KickOffAlly {
         tools_data: &mut ToolData,
         action_wrapper: &mut ActionWrapper,
     ) -> bool {
-        action_wrapper.push(0, MoveTo::new(Point2::new(-0.25, 2.5), -PI / 4.0, 0.0,None));
-        action_wrapper.push(1, MoveTo::new(Point2::new(-1.5, -1.5), -PI / 4.0, 0.0,None));
-        action_wrapper.push(2, MoveTo::new(Point2::new(-0.25, -2.5), -PI / 4.0, 0.0,None));
-        action_wrapper.push(3, MoveTo::new(Point2::new(-0.2, 0.0), -PI / 4.0, 0.0,None));
-        action_wrapper.push(4, MoveTo::new(Point2::new(-1.5, 1.5), -PI / 4.0, 0.0,None));
+        action_wrapper.push(0, MoveTo::new(Point2::new(-1.5, 1.1), -PI / 4.0, 0.0,None));
+        action_wrapper.push(1, MoveTo::new(Point2::new(-1.0, -0.0), -PI / 4.0, 0.0,None));
+        action_wrapper.push(2, MoveTo::new(Point2::new(-1.5, -1.1), -PI / 4.0, 0.0,None));
+        action_wrapper.push(3, MoveTo::new(Point2::new(-1.2, 0.7), -PI / 4.0, 0.0,None));
+        action_wrapper.push(4, MoveTo::new(Point2::new(-1.6, -0.7), -PI / 4.0, 0.0,None));
         action_wrapper.push(5, MoveTo::new(Point2::new(-4.0, -0.0), -PI / 4.0, 0.0,None));
         true
     }
 }
+
